prost-types: Extract modules to separate files.
